--- src/docs/asciidoc/api-doc-admin.adoc
@@ -1,41 +1,0 @@
-= Admin API
-:toc: left
-:toclevels: 2
-:source-highlighter: highlightjs
-
-== 0. 관리자 페이지
-=== 0-1. 카페 저장(Post) - 성공
-==== Http request
-include::{snippets}/admin/cafe/save/success/http-request.adoc[]
-==== Request parts
-include::{snippets}/admin/cafe/save/success/request-parts.adoc[]
-==== Request part request fields
-include::{snippets}/admin/cafe/save/success/request-part-request-fields.adoc[]
-==== Http response
-include::{snippets}/admin/cafe/save/success/http-response.adoc[]
-==== Response fields
-include::{snippets}/admin/cafe/save/success/response-fields.adoc[]
-
-=== 0-2. 카페 조회(존재하면 카페 정보 반환) - 성공
-==== Http request
-include::{snippets}/admin/cafe/success/http-request.adoc[]
-==== Query parameters
-include::{snippets}/admin/cafe/success/query-parameters.adoc[]
-==== Http response
-include::{snippets}/admin/cafe/success/http-response.adoc[]
-==== Response fields
-include::{snippets}/admin/cafe/success/response-fields.adoc[]
-
-=== 0-3. 카페 존재 여부 조회(Get) - 성공
-==== Http request
-include::{snippets}/admin/cafe/exist/success/http-request.adoc[]
-==== Query parameters
-include::{snippets}/admin/cafe/exist/success/query-parameters.adoc[]
-==== Http response
-include::{snippets}/admin/cafe/exist/success/http-response.adoc[]
-==== Response fields
-include::{snippets}/admin/cafe/exist/success/response-fields.adoc[]
-
-
-=== link:api-docs.html[API 목록으로]
-
