+++ src/docs/asciidoc/api-docs.adoc
@@ -2,7 +2,6 @@
 :toc: left
 :toclevels: 2
 
-=== link:api-doc-admin.html[0. 관리자(Admin) API]
 === link:api-doc-bookmark.html[1. 북마크(Bookmark) API]
 === link:api-doc-bookmark-folder.html[2. 북마크 폴더(Bookmark Folder) API]
 === link:api-doc-plan.html[3. 계획하기(Plan) API]
