+++ src/main/java/com/sideproject/cafe_cok/admin/application/AdminService.java
@@ -1,12 +1,11 @@
 package com.sideproject.cafe_cok.admin.application;
 
-
+import com.sideproject.cafe_cok.admin.dto.*;
 import com.sideproject.cafe_cok.admin.dto.request.AdminCafeSaveRequest;
-import com.sideproject.cafe_cok.admin.dto.request.AdminMenuSaveRequest;
-import com.sideproject.cafe_cok.admin.dto.response.AdminCafeExistResponse;
-import com.sideproject.cafe_cok.admin.dto.response.AdminCafeSaveResponse;
-import com.sideproject.cafe_cok.admin.dto.response.AdminCafeFindResponse;
+import com.sideproject.cafe_cok.admin.dto.request.AdminCafeUpdateRequest;
+import com.sideproject.cafe_cok.admin.dto.request.AdminMenuRequestDto;
 import com.sideproject.cafe_cok.admin.dto.response.AdminRestoreMemberResponse;
+import com.sideproject.cafe_cok.admin.dto.response.AdminSuccessAndRedirectResponse;
 import com.sideproject.cafe_cok.admin.exception.NoWithdrawalMemberException;
 import com.sideproject.cafe_cok.cafe.domain.Cafe;
 import com.sideproject.cafe_cok.cafe.domain.OperationHour;
@@ -15,28 +14,29 @@ import com.sideproject.cafe_cok.cafe.domain.repository.OperationHourRepository;
 import com.sideproject.cafe_cok.image.domain.Image;
 import com.sideproject.cafe_cok.image.domain.enums.ImageType;
 import com.sideproject.cafe_cok.image.domain.repository.ImageRepository;
-import com.sideproject.cafe_cok.image.dto.CafeMainImageDto;
-import com.sideproject.cafe_cok.image.dto.CafeOtherImageDto;
+import com.sideproject.cafe_cok.image.dto.ImageDto;
+import com.sideproject.cafe_cok.member.domain.Feedback;
 import com.sideproject.cafe_cok.member.domain.Member;
+import com.sideproject.cafe_cok.member.domain.enums.FeedbackCategory;
+import com.sideproject.cafe_cok.member.domain.repository.FeedbackRepository;
 import com.sideproject.cafe_cok.member.domain.repository.MemberRepository;
 import com.sideproject.cafe_cok.menu.domain.Menu;
 import com.sideproject.cafe_cok.menu.domain.repository.MenuRepository;
-import com.sideproject.cafe_cok.menu.dto.CafeSaveMenuDto;
 import com.sideproject.cafe_cok.review.domain.Review;
 import com.sideproject.cafe_cok.utils.FormatConverter;
 import com.sideproject.cafe_cok.utils.S3.component.S3Uploader;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
-import java.math.BigDecimal;
+import java.time.DayOfWeek;
 import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import static com.sideproject.cafe_cok.utils.Constants.*;
 import static com.sideproject.cafe_cok.utils.FormatConverter.*;
@@ -47,63 +47,217 @@ import static com.sideproject.cafe_cok.utils.FormatConverter.*;
 public class AdminService {
 
     private final MemberRepository memberRepository;
+    private final FeedbackRepository feedbackRepository;
     private final CafeRepository cafeRepository;
     private final MenuRepository menuRepository;
     private final ImageRepository imageRepository;
     private final OperationHourRepository operationHourRepository;
-
-
     private final S3Uploader s3Uploader;
 
-    private final Integer START_TIME_IDX = 0;
-    private final Integer END_TIME_IDX = 1;
-    private final Boolean CAFE_IS_NOT_CLOSED = false;
-    private final Boolean CAFE_IS_CLOSED = true;
+    public List<AdminSuggestionDto> findFeedbackByCategory(final FeedbackCategory feedbackCategory) {
+
+        List<Feedback> findFeedbacks = feedbackRepository.findByCategoryOrderByIdDesc(feedbackCategory);
+        return findFeedbacks.stream()
+                .map(feedback -> new AdminSuggestionDto(feedback))
+                .collect(Collectors.toList());
+    }
+
+    public List<AdminCafeDto> findCafes() {
+
+        List<Cafe> findCafes = cafeRepository.findAllByOrderByIdDesc();
+        return findCafes.stream()
+                .map(cafe -> new AdminCafeDto(cafe))
+                .collect(Collectors.toList());
+    }
+
+    public AdminCafeDto findCafeById(final Long id) {
+        Cafe findCafe = cafeRepository.getById(id);
+        List<Image> filteredImage = findCafe.getImages().stream()
+                .filter(image -> image.getImageType().equals(ImageType.CAFE) || image.getImageType().equals(ImageType.CAFE_MAIN))
+                .collect(Collectors.toList());
+        List<ImageDto> images = ImageDto.fromList(filteredImage);
+
+        List<AdminMenuDto> findMenus = menuRepository.findByCafeId(id).stream()
+                .map(menu -> {
+                    List<Image> findMenuImages = imageRepository.findByMenu(menu);
+                    if (findMenuImages.isEmpty()) return new AdminMenuDto(menu);
+                    return new AdminMenuDto(menu, ImageDto.from(findMenuImages.get(0)));
+                }).collect(Collectors.toList());
 
 
-    @Transactional
-    public AdminCafeSaveResponse saveCafe(final AdminCafeSaveRequest request,
-                                          final MultipartFile mainImage,
-                                          final List<MultipartFile> otherImages) {
-
-        Cafe cafe = request.toEntity();
-        Optional<Cafe> findOptionalCafe = cafeRepository.findByLatitudeAndLongitude(request.getMapy(), request.getMapx());
-        if(findOptionalCafe.isPresent()) {
-
-            cafe = findOptionalCafe.get();
-            List<Image> findImages = imageRepository.findImageByCafe(cafe);
-            List<Menu> findMenus = menuRepository.findByCafeId(cafe.getId());
-
-            for (Menu menu : findMenus) {
-                List<Image> findMenuImages = imageRepository.findByMenu(menu);
-                if(findMenuImages.isEmpty()) continue;
-                findImages.add(findImages.get(0));
+        List<OperationHour> findOperationHourList = operationHourRepository.findByCafeId(findCafe.getId());
+        List<AdminOperationHourDto> hours = generateHours();
+        for (AdminOperationHourDto hour : hours) {
+            String day = hour.getDay();
+            for (OperationHour operationHour : findOperationHourList) {
+                if(day.equals(FormatConverter.getKoreanDayOfWeek(operationHour.getDate()))) {
+                    hour.changeTime(operationHour);
+                }
             }
-
-            for (Image findImage : findImages) {
-                String origin = findImage.getOrigin();
-                String thumbnail = findImage.getThumbnail();
-                String medium = findImage.getMedium();
-
-                if(origin != null) s3Uploader.delete(origin);
-                if(thumbnail != null) s3Uploader.delete(origin);
-                if(medium != null) s3Uploader.delete(origin);
-            }
-
-            imageRepository.deleteAll(findImages);
-            menuRepository.deleteAll(findMenus);
-            List<OperationHour> findOperationHours = operationHourRepository.findByCafeId(cafe.getId());
-            operationHourRepository.deleteAll(findOperationHours);
-            cafe.changeCafe(request.getName(), request.getTelephone(), request.getRoadAddress(), request.getMapx(), request.getMapy());
         }
 
-        Cafe savedCafe = cafeRepository.save(cafe);
-        CafeMainImageDto mainImageDto = saveCafeMainImages(mainImage, savedCafe);
-        List<CafeOtherImageDto> otherImageDtos = saveCafeOtherImages(otherImages, savedCafe);
-        List<CafeSaveMenuDto> menuDtos = saveMenu(request.getMenus(), savedCafe);
-        List<List<String>> operationHours = saveOperationHours(savedCafe, request.getHours());
+        return new AdminCafeDto(findCafe, images, findMenus, hours);
+    }
 
-        return AdminCafeSaveResponse.of(savedCafe, mainImageDto, otherImageDtos, menuDtos, operationHours);
+    @Transactional
+    public AdminSuccessAndRedirectResponse updateCafe(final Long id,
+                                                      final AdminCafeUpdateRequest request) {
+
+        Cafe findCafe = cafeRepository.getById(id);
+        findCafe.setName(request.getName());
+        findCafe.setRoadAddress(request.getAddress());
+        findCafe.setPhoneNumber(request.getPhoneNumber());
+        cafeRepository.save(findCafe);
+
+
+        //카페 메인 이미지 수정
+        if(request.getImage().getImageBase64() != null) {
+
+            File converted = convertBase64StringToFile(request.getImage().getImageBase64());
+            String originImageUrl = s3Uploader.upload(converted, CAFE_MAIN_ORIGIN_IMAGE_DIR);
+
+            Optional<Image> optionalImage = imageRepository.findById(request.getImage().getId());
+            if(optionalImage.isPresent()) {
+                Image findImage = optionalImage.get();
+                s3Uploader.delete(findImage.getOrigin());
+                s3Uploader.delete(findImage.getMedium());
+                s3Uploader.delete(findImage.getThumbnail());
+                findImage.changeOrigin(originImageUrl);
+                findImage.changMedium(changePath(originImageUrl, CAFE_MAIN_ORIGIN_IMAGE_DIR, CAFE_MAIN_MEDIUM_IMAGE_DIR));
+                findImage.changeThumbnail(changePath(originImageUrl, CAFE_MAIN_ORIGIN_IMAGE_DIR, CAFE_MAIN_THUMBNAIL_DIR));
+                imageRepository.save(findImage);
+            }
+        }
+
+        //카페 나머지 이미지 수정
+        for (AdminImageDto otherImage : request.getOtherImages()) {
+
+            File converted = convertBase64StringToFile(otherImage.getImageBase64());
+            String originImageUrl = s3Uploader.upload(converted, CAFE_ORIGIN_IMAGE_DIR);
+            String thumbnailImageUrl = changePath(originImageUrl, CAFE_ORIGIN_IMAGE_DIR, CAFE_THUMBNAIL_IMAGE_DIR);
+            //기존 이미지가 수정되는 경우
+            if(otherImage.getId() != null) {
+                Optional<Image> optionalImage = imageRepository.findById(otherImage.getId());
+                if(optionalImage.isPresent()) {
+                    Image findImage = optionalImage.get();
+                    s3Uploader.delete(findImage.getThumbnail());
+                    s3Uploader.delete(findImage.getOrigin());
+                    findImage.changeOrigin(originImageUrl);
+                    findImage.changeThumbnail(thumbnailImageUrl);
+                    imageRepository.save(findImage);
+                }
+                continue;
+            }
+
+            //새로운 이미지가 추가된 경우
+            Image newImage = new Image(ImageType.CAFE, originImageUrl, thumbnailImageUrl, findCafe);
+            imageRepository.save(newImage);
+        }
+
+        //운영 시간 수정
+        List<AdminOperationHourDto> hours = request.getHours();
+        operationHourRepository.deleteByCafeId(id);
+        if(checkoutInputHours(hours)) saveOperationHours(hours, findCafe);
+
+        //메뉴 수정
+        List<AdminMenuRequestDto> menus = request.getMenus();
+        for (AdminMenuRequestDto menu : menus) {
+
+            //기존 메뉴인 경우
+            Menu targetMenu;
+            if(menu.getId() != null) {
+                Optional<Menu> optionalMenu = menuRepository.findById(menu.getId());
+                if(optionalMenu.isEmpty()) continue;
+                targetMenu = optionalMenu.get();
+                targetMenu.changeName(menu.getName());
+                targetMenu.changePrice(menu.getPrice());
+            }
+            else  targetMenu = new Menu(menu.getName(), menu.getPrice(), findCafe);
+            menuRepository.save(targetMenu);
+
+            //변경하려는 이미지가 존재하는 경우
+            if (menu.getImage() != null && !menu.getImage().isEmpty()) {
+
+                File converted = convertBase64StringToFile(menu.getImage());
+                String originImageUrl = s3Uploader.upload(converted, MENU_ORIGIN_IMAGE_DIR);
+                String thumbnailImageUrl = changePath(originImageUrl, MENU_ORIGIN_IMAGE_DIR, MENU_THUMBNAIL_IMAGE_DIR);
+
+                List<Image> findImages = imageRepository.findByMenu(targetMenu);
+                Image targetImage;
+                if(findImages.isEmpty()) {
+                    targetImage = new Image(ImageType.MENU, originImageUrl, thumbnailImageUrl, findCafe, targetMenu);
+                } else {
+                    targetImage = findImages.get(0);
+                    targetImage.changeOrigin(originImageUrl);
+                    targetImage.changeThumbnail(thumbnailImageUrl);
+                }
+
+                imageRepository.save(targetImage);
+            }
+        }
+
+        return new AdminSuccessAndRedirectResponse("Update successful", "/admin/cafe/" + id);
+    }
+
+    @Transactional
+    public AdminSuccessAndRedirectResponse saveCafe(final AdminCafeSaveRequest request) {
+
+        Cafe newCafe = new Cafe(request);
+        Cafe savedCafe = cafeRepository.save(newCafe);
+
+        //메인 이미지 저장
+        File converted = convertBase64StringToFile(request.getMainImage());
+        String originImageUrl = s3Uploader.upload(converted, CAFE_MAIN_ORIGIN_IMAGE_DIR);
+        String thumbnailImageDir = changePath(originImageUrl, CAFE_MAIN_ORIGIN_IMAGE_DIR, CAFE_MAIN_THUMBNAIL_DIR);
+        String midImageDir = changePath(originImageUrl, CAFE_MAIN_ORIGIN_IMAGE_DIR, CAFE_MAIN_MEDIUM_IMAGE_DIR);
+        Image mainImage = new Image(ImageType.CAFE_MAIN, originImageUrl, thumbnailImageDir, midImageDir, savedCafe);
+        imageRepository.save(mainImage);
+
+        //카페 나머지 이미지 저장
+        for (String otherImage : request.getOtherImages()) {
+            converted = convertBase64StringToFile(otherImage);
+            originImageUrl = s3Uploader.upload(converted, CAFE_ORIGIN_IMAGE_DIR);
+            thumbnailImageDir = changePath(originImageUrl, CAFE_ORIGIN_IMAGE_DIR, CAFE_THUMBNAIL_IMAGE_DIR);
+            Image othreImage = new Image(ImageType.CAFE, originImageUrl, thumbnailImageDir, savedCafe);
+            imageRepository.save(othreImage);
+        }
+
+        //메뉴 저장
+        List<AdminMenuRequestDto> menus = request.getMenus();
+        for (AdminMenuRequestDto menu : menus) {
+            Menu newMenu = new Menu(menu.getName(), menu.getPrice(), savedCafe);
+            menuRepository.save(newMenu);
+
+            //메뉴 이미지 저장
+            converted = convertBase64StringToFile(menu.getImage());
+            originImageUrl = s3Uploader.upload(converted, MENU_ORIGIN_IMAGE_DIR);
+            thumbnailImageDir = changePath(originImageUrl, MENU_ORIGIN_IMAGE_DIR, MENU_THUMBNAIL_IMAGE_DIR);
+            Image menuImage = new Image(ImageType.MENU, originImageUrl, thumbnailImageDir, savedCafe);
+            imageRepository.save(menuImage);
+        }
+
+        //운영시간 저장
+        List<AdminOperationHourDto> hours = request.getHours();
+        if(checkoutInputHours(hours)) saveOperationHours(hours, savedCafe);
+
+        return new AdminSuccessAndRedirectResponse("Update successful", "/admin/cafe/" + savedCafe.getId());
+    }
+
+    @Transactional
+    public boolean menuDelete(final Long id) {
+
+        Optional<Menu> optionalMenu = menuRepository.findById(id);
+        if(optionalMenu.isEmpty()) return true;
+
+        Menu findMenu = optionalMenu.get();
+        List<Image> findMenuImages = imageRepository.findByMenu(findMenu);
+        for (Image findMenuImage : findMenuImages) {
+            s3Uploader.delete(findMenuImage.getThumbnail());
+            s3Uploader.delete(findMenuImage.getOrigin());
+        }
+        imageRepository.deleteAll(findMenuImages);
+        menuRepository.delete(findMenu);
+        return true;
     }
 
     @Transactional
@@ -121,142 +275,42 @@ public class AdminService {
         return new AdminRestoreMemberResponse(findMember);
     }
 
-    public AdminCafeExistResponse checkCafeExist(final BigDecimal mapx,
-                                                 final BigDecimal mapy) {
+    private boolean checkoutInputHours(List<AdminOperationHourDto> hours) {
 
-        Optional<Cafe> findOptionalCafe = cafeRepository.findByLatitudeAndLongitude(mapy, mapx);
-        if(findOptionalCafe.isEmpty()) return new AdminCafeExistResponse(false);
-        return new AdminCafeExistResponse(true);
-    }
-
-    public AdminCafeFindResponse findCafe(final BigDecimal mapx,
-                                          final BigDecimal mapy) {
-
-        Optional<Cafe> findOptionalCafe = cafeRepository.findByLatitudeAndLongitude(mapy, mapx);
-        if(findOptionalCafe.isEmpty()) return new AdminCafeFindResponse();
-
-        Cafe findCafe = findOptionalCafe.get();
-        List<Image> findCafeImages = imageRepository.findImageByCafe(findCafe);
-        CafeMainImageDto findCafeMainImage = null;
-        List<CafeOtherImageDto> findCafeOtherImages = new ArrayList<>();
-        for (Image findCafeImage : findCafeImages) {
-            if(findCafeImage.getImageType().equals(ImageType.CAFE_MAIN)) {
-                findCafeMainImage = CafeMainImageDto.from(findCafeImage);
-                continue;
-            }
-            if(findCafeImage.getImageType().equals(ImageType.CAFE)) {
-                findCafeOtherImages.add(CafeOtherImageDto.from(findCafeImage));
-            }
+        for (AdminOperationHourDto hour : hours) {
+            if(hour.getStartHour() != 0) return true;
+            if(hour.getStartMinute() != 0) return true;
+            if(hour.getEndHour() != 0) return true;
+            if(hour.getEndMinute() != 0) return true;
         }
 
-        List<CafeSaveMenuDto> findMenuDtoList = new ArrayList<>();
-        List<Menu> findMenus = menuRepository.findByCafeId(findCafe.getId());
-        for (Menu findMenu : findMenus) {
-            List<Image> findMenuImages = imageRepository.findByMenu(findMenu);
-            if(findMenuImages.isEmpty()) findMenuDtoList.add(CafeSaveMenuDto.from(findMenu));
-            for (Image findMenuImage : findMenuImages) {
-                findMenuDtoList.add(CafeSaveMenuDto.of(findMenu, findMenuImage));
-            }
+        return false;
+    }
+
+    private List<AdminOperationHourDto> generateHours() {
+
+        List<String> daysOfWeek = Arrays.asList("월", "화", "수", "목", "금", "토", "일");
+        List<AdminOperationHourDto> hours = new ArrayList<>();
+        for (String day : daysOfWeek) {
+            hours.add(new AdminOperationHourDto(day));
         }
 
-        List<OperationHour> findOperationHourList = operationHourRepository.findByCafeId(findCafe.getId());
-        List<List<String>> findHours = convertOperationHoursToListString(findOperationHourList);
-        AdminCafeSaveResponse cafe = AdminCafeSaveResponse
-                .of(findCafe, findCafeMainImage, findCafeOtherImages, findMenuDtoList, findHours);
-
-        return new AdminCafeFindResponse(cafe);
+        return hours;
     }
 
-    private CafeMainImageDto saveCafeMainImages(final MultipartFile mainImage, final Cafe cafe) {
-        String cafeMainOriginImageUrl = s3Uploader.upload(mainImage, CAFE_MAIN_ORIGIN_IMAGE_DIR);
-        Image cafeMainImage = new Image(
-                ImageType.CAFE_MAIN
-                , cafeMainOriginImageUrl
-                , changePath(cafeMainOriginImageUrl, CAFE_MAIN_ORIGIN_IMAGE_DIR, CAFE_MAIN_THUMBNAIL_DIR)
-                , changePath(cafeMainOriginImageUrl, CAFE_MAIN_ORIGIN_IMAGE_DIR, CAFE_MAIN_MEDIUM_IMAGE_DIR)
-                , cafe);
+    private void saveOperationHours(final List<AdminOperationHourDto> hours, final Cafe cafe) {
+        List<OperationHour> newOperationHours = new ArrayList<>();
+        for (AdminOperationHourDto hour : hours) {
+            DayOfWeek day = getDyaOfWeekByKoreanDay(hour.getDay());
+            LocalTime startTime = LocalTime.of(hour.getStartHour(), hour.getStartMinute());
+            LocalTime endTime = LocalTime.of(hour.getEndHour(), hour.getEndMinute());
+            boolean isClosed = false;
 
-        Image savedImage = imageRepository.save(cafeMainImage);
-        return CafeMainImageDto.from(savedImage);
-    }
-
-    private List<CafeOtherImageDto> saveCafeOtherImages(final List<MultipartFile> otherImages, final Cafe cafe) {
-
-        List<CafeOtherImageDto> cafeOtherImageDtos = new ArrayList<>();
-        if(otherImages != null) {
-            for (MultipartFile otherImage : otherImages) {
-                String otherImageOriginUrl = s3Uploader.upload(otherImage, CAFE_ORIGIN_IMAGE_DIR);
-                Image cafeImage = new Image(ImageType.CAFE
-                        , otherImageOriginUrl
-                        , changePath(otherImageOriginUrl, CAFE_ORIGIN_IMAGE_DIR, CAFE_THUMBNAIL_IMAGE_DIR)
-                        , cafe);
-                Image savedImage = imageRepository.save(cafeImage);
-                cafeOtherImageDtos.add(CafeOtherImageDto.from(savedImage));
-            }
+            if(startTime.equals(LocalTime.MIDNIGHT) && endTime.equals(LocalTime.MIDNIGHT)) isClosed = true;
+            OperationHour newOperationHour = new OperationHour(day, startTime, endTime, isClosed, cafe);
+            newOperationHours.add(newOperationHour);
         }
 
-        return cafeOtherImageDtos;
-    }
-
-    private List<CafeSaveMenuDto> saveMenu(final List<AdminMenuSaveRequest> menus, final Cafe cafe) {
-        List<CafeSaveMenuDto> menuDtos = new ArrayList<>();
-
-        if(menus.isEmpty()) return menuDtos;
-
-        for (AdminMenuSaveRequest menu : menus) {
-            Menu savedMenu = menuRepository.save(menu.toEntity(cafe));
-
-            if (menu.getImage() != null && !menu.getImage().isEmpty()) {
-
-                File convertedFile = convertBase64StringToFile(menu.getImage());
-                String originMenuUrl = s3Uploader.upload(convertedFile, MENU_ORIGIN_IMAGE_DIR);
-                Image menuImage = new Image(ImageType.MENU
-                        , originMenuUrl
-                        , changePath(originMenuUrl, MENU_ORIGIN_IMAGE_DIR, MENU_THUMBNAIL_IMAGE_DIR)
-                        , cafe
-                        , savedMenu);
-                Image saveMenuImage = imageRepository.save(menuImage);
-                menuDtos.add(CafeSaveMenuDto.of(savedMenu, saveMenuImage));
-                continue;
-            }
-            menuDtos.add(CafeSaveMenuDto.from(savedMenu));
-        }
-        return menuDtos;
-    }
-
-    private List<List<String>> saveOperationHours(final Cafe cafe, List<List<String>> hours) {
-
-        List<OperationHour> operationHours = new ArrayList<>();
-
-        if(allOperatingHoursEmpty(hours)) return createEmptyOperationHours();
-
-        for (int i = 0; i < hours.size(); i++) {
-            List<String> hour = hours.get(i);
-
-            LocalTime startTime = LocalTime.of(0, 0, 0);
-            LocalTime endTime = LocalTime.of(0, 0, 0);
-            boolean cafeIsClosed = CAFE_IS_CLOSED;
-
-            if(!hour.get(START_TIME_IDX).isEmpty()) {
-                startTime = LocalTime.parse(hour.get(START_TIME_IDX));
-                endTime = LocalTime.parse(hour.get(END_TIME_IDX));
-                cafeIsClosed = CAFE_IS_NOT_CLOSED;
-            }
-            operationHours.add(new OperationHour(operationDays.get(i), startTime, endTime, cafeIsClosed, cafe));
-        }
-
-        List<OperationHour> savedOperationHours = operationHourRepository.saveAll(operationHours);
-        return FormatConverter.convertOperationHoursToListString(savedOperationHours);
-    }
-
-    private boolean allOperatingHoursEmpty(final List<List<String>> hours) {
-        for (List<String> hour : hours) if(!hour.get(START_TIME_IDX).isEmpty()) return false;
-        return true;
-    }
-
-    private List<List<String>> createEmptyOperationHours() {
-        List<List<String>> emptyHours = new ArrayList<>();
-        for(int i = 0; i < 7; i++) emptyHours.add(Arrays.asList("", ""));
-        return emptyHours;
+        operationHourRepository.saveAll(newOperationHours);
     }
 }
