+++ src/main/resources/static/css/cafe-style.css
@@ -1,0 +1,39 @@
+.btn-container {
+    display: flex;
+    gap: 5px;
+}
+
+.submit-button-container {
+    position: absolute;
+    right: 10px;
+    bottom: 80px;
+}
+
+.submit-button-container button {
+    font-size: 1.25rem;
+    padding: 5px 30px;
+}
+
+.checkbox-container {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 10px;
+}
+.form-check {
+    margin: 0;
+}
+.form-check-input {
+    transform: scale(1.2);
+}
+.form-check-label {
+    font-size: 1.2rem;
+    padding: 5px 10px;
+    background-color: black;
+    color: white;
+    font-weight: bold;
+    border-radius: 5px;
+    transition: background-color 0.3s;
+}
+.form-check-label:hover {
+    background-color: #e9ecef;
+}
