+++ src/main/resources/templates/fragments/sidebar.html
@@ -7,6 +7,10 @@
                 <div class="sb-nav-link-icon"><i class="fas fa-columns"></i></div>
                 카페 목록
             </a>
+            <a class="nav-link" href="/admin/cafe/register">
+                <div class="sb-nav-link-icon"><i class="fas fa-columns"></i></div>
+                카페 등록
+            </a>
             <div class="sb-sidenav-menu-heading">피드백</div>
             <a class="nav-link" href="/admin/suggestions">
                 <div class="sb-nav-link-icon"><i class="fas fa-table"></i></div>
