+++ src/main/resources/templates/layouts/default_layout.html
@@ -11,24 +11,22 @@
     <script src="https://use.fontawesome.com/releases/v6.1.0/js/all.js" crossorigin="anonymous"></script>
 </head>
 <body class="sb-nav-fixed">
-<th:block th:replace="fragments/header :: headerFragment"></th:block>
+<th:block th:replace="~{fragments/header :: headerFragment}"></th:block>
 
 <div id="layoutSidenav">
     <div id="layoutSidenav_nav">
-        <th:block th:replace="fragments/sidebar :: sidebarFragment"></th:block>
+        <th:block th:replace="~{fragments/sidebar :: sidebarFragment}"></th:block>
     </div>
     <div id="layoutSidenav_content">
         <main>
             <th:block layout:fragment="content"></th:block>
         </main>
-        <th:block th:replace="fragments/footer :: footerFragment"></th:block>
+        <th:block th:replace="~{fragments/footer :: footerFragment}"></th:block>
     </div>
 </div>
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" crossorigin="anonymous"></script>
 <script th:src="@{/js/scripts.js}"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.8.0/Chart.min.js" crossorigin="anonymous"></script>
-<!--<script th:src="@{/assets/demo/chart-area-demo.js}"></script>-->
-<!--<script th:src="@{/assets/demo/chart-bar-demo.js}"></script>-->
 <script src="https://cdn.jsdelivr.net/npm/simple-datatables@latest" crossorigin="anonymous"></script>
 <script th:src="@{/js/datatables-simple-demo.js}"></script>
 </body>
